pyproject.toml
[project]
name = "sepapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.3",
    "django>=5.2.8",
    "django-allauth>=65.13.1",
    "django-widget-tweaks==1.5.0",
    "pillow>=12.0.0",
    "pyjwt>=2.10.1",
    "python-decouple==3.8",
    "requests==2.31.0",
]
